work in progress

+++ subprojects/nxf-commons/src/main/nextflow/util/MemoryUnit.groovy
@@ -118,7 +118,7 @@ class MemoryUnit implements Comparable<MemoryUnit>, Serializable, Cloneable {
     }
 
     def multiply( Number value ) {
-        return new MemoryUnit( size * value )
+        return new MemoryUnit( size * value.toLong() )
     }
 
     def div( Number value ) {
